Remove unused code & simplify glib.go
